Move helpers to lib
