+++ Frontend/src/pages/schedule/TableView.jsx
@@ -1,6 +1,11 @@
 import { getFormatName } from '@wca/helpers'
 import React, { useState } from 'react'
 import { Checkbox, Header, Segment, Table, TableCell } from 'semantic-ui-react'
+import {
+  earliestWithLongestTieBreaker,
+  groupActivities,
+} from '../../lib/activities'
+import { activitiesByDate, getLongDate, getShortTime } from '../../lib/dates'
 
 export default function TableView({ dates, timeZone, venuesShown, events }) {
   const rounds = events.flatMap((event) => event.rounds)
@@ -147,75 +152,5 @@ function ActivityRow({ isExpanded, activityGroup, round, allRooms, timeZone }) {
         </>
       )}
     </Table.Row>
-  )
-}
-
-// TODO: move to separate utils file
-
-const earliestWithLongestTieBreaker = (a, b) => {
-  if (a.startTime < b.startTime) {
-    return -1
-  }
-  if (a.startTime > b.startTime) {
-    return 1
-  }
-  if (a.endTime < b.endTime) {
-    return 1
-  }
-  if (a.endTime > b.endTime) {
-    return -1
-  }
-  return 0
-}
-
-const activitiesByDate = (activities, date, timeZone) => {
-  return activities.filter(
-    // not sure how to check startTime *in timeZone* is on date, besides
-    // comparing locale strings, which seems bad
-    // (there's only .getDay() for local time zone and .getUTCDay() for UTC,
-    // but no such function for arbitrary time zone)
-    (activity) =>
-      new Date(activity.startTime).toLocaleDateString([], { timeZone }) ===
-      date.toLocaleDateString()
-  )
-}
-
-const getShortTime = (date, timeZone) => {
-  return new Date(date).toLocaleTimeString([], {
-    timeStyle: 'short',
-    timeZone,
-  })
-}
-
-const getLongDate = (date) => {
-  return new Date(date).toLocaleDateString([], {
-    weekday: 'long',
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-  })
-}
-
-// assumes they are sorted
-const groupActivities = (activities) => {
-  const grouped = []
-  activities.forEach((activity) => {
-    if (
-      grouped.length > 0 &&
-      areGroupable(activity, grouped[grouped.length - 1][0])
-    ) {
-      grouped[grouped.length - 1].push(activity)
-    } else {
-      grouped.push([activity])
-    }
-  })
-  return grouped
-}
-
-const areGroupable = (act1, act2) => {
-  return (
-    act1.startTime === act2.startTime &&
-    act1.endTime === act2.endTime &&
-    act1.activityCode === act2.activityCode
   )
 }
